pyproject.toml
[project]
name = "llm-ocr-backend"
version = "0.1.0"
description = "FastAPI + LangChain + DeepSeek-OCR backend for autonomous market research agent"
readme = "README.md"
authors = [{name = "Your Name", email = "your@email.com"}]
license = {text = "MIT"}
keywords = ["langgraph", "rag", "ocr", "fastapi", "ai-agent"]
requires-python = ">=3.10"

dependencies = [
    # Web / API
    "fastapi",
    "uvicorn[standard]",
    # LangChain ecosystem (pin major versions for stability)
    "langchain >=1.1.0",
    "langchain-community >=0.4.0",
    "langchain-google-genai >=4.0.0",
    "langchain-huggingface >=1.2.0",
    "langgraph >=1.0.5",
    "langchain-chroma >=1.1.0",
    # Vector DB / Tools
    "chromadb >=1.3.0",
    "tavily-python",
    "beautifulsoup4",
    "python-dotenv",
    "tiktoken",
    # Vision / DeepSeek-OCR
    "torch >=2.0.0",
    "transformers >=4.46.3",
    "pillow >=10.0.0",
    "accelerate >=0.26.0",
    "safetensors >=0.4.2",
    "timm >=0.9.12",
    "einops >=0.7.0",
    "sentencepiece >=0.2.0",
    "opencv-python",
    # Stability / Conflicts
    "protobuf <5", 
    "pydantic >=2.0,<3",
    # PDF processing (required for uploaded PDFs → images for DeepSeek-OCR)
    "pdf2image>=1.17.0",
    # Embeddings 
    "sentence-transformers >=5.2.0",
    "aiofiles==25.1.0",
    "IPython"

]

[project.optional-dependencies]
dev = [
    "ipykernel",
    "black",
    "ruff",
    "mypy",
    "rich",
    "loguru"
]

[tool.setuptools]
packages = ["app"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"